# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BDPM Drug Resolver",
    "description": "Détection, recherche et correction de noms de médicaments français à partir de la Base de Données Publique des Médicaments (BDPM). Correspondance lexicale locale, sans dépendance externe au runtime — ne constitue pas une validation médicale ou pharmaceutique.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/update-database": {
      "post": {
        "summary": "Update Database",
        "operationId": "update_database_admin_update_database_post",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "AdminKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip_download",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Skip Download"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDatabaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateDatabaseResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "drug_count": {
            "type": "integer",
            "title": "Drug Count"
          },
          "alias_count": {
            "type": "integer",
            "title": "Alias Count"
          },
          "rejected_ratio": {
            "type": "number",
            "title": "Rejected Ratio"
          }
        },
        "type": "object",
        "required": [
          "status",
          "drug_count",
          "alias_count",
          "rejected_ratio"
        ],
        "title": "UpdateDatabaseResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AdminKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      }
    }
  }
}